gradle: Refactor and remove old deprecated properties and methods
